Refactoried
