--- nvim/lua/config/interface.lua
@@ -1,320 +1,0 @@
-local vim = vim
-
-require'nvim-web-devicons'.setup {
- -- your personnal icons can go here (to override)
- -- DevIcon will be appended to `name`
- override = {
-  zsh = {
-    icon = "",
-    color = "#428850",
-    name = "Zsh"
-  }
- };
- -- globally enable default icons (default to false)
- -- will get overriden by `get_icons` option
- default = true;
-}
--- Nvim-Tree config
-vim.g.nvim_tree_side = 'left' -- by default
-vim.g.nvim_tree_width =30 -- by default, can be width_in_columns or 'width_in_percent%'
-vim.g.nvim_tree_ignore = {} -- by default
-vim.g.nvim_tree_gitignore = 0 -- by default
-vim.g.nvim_tree_auto_open = 1 -- by default, opens the tree when typing `vim $DIR` or `vim`
-vim.g.nvim_tree_auto_close = 1 -- by default, closes the tree when it's the last window
-vim.g.nvim_tree_auto_ignore_ft = { 'dashboard' } -- by default, don't auto open tree on specific filetypes.
-vim.g.nvim_tree_quit_on_open = 0 -- by default, closes the tree when you open a file
-vim.g.nvim_tree_follow = 1 -- by default, this option allows the cursor to be updated when entering a buffer
-vim.g.nvim_tree_indent_markers = 1 -- by default, this option shows indent markers when folders are open
-vim.g.nvim_tree_hide_dotfiles = 1 -- by default, this option hides files and folders starting with a dot `.`
-vim.g.nvim_tree_git_hl = 1 -- by default, will enable file highlight for git attributes (can be used without the icons).
-vim.g.nvim_tree_highlight_opened_files = 1 -- by default, will enable folder and file icon highlight for opened files/directories.
-vim.g.nvim_tree_root_folder_modifier = ':~' -- is the default. See :help filename-modifiers for more options
-vim.g.nvim_tree_tab_open = 1 -- by default, will open the tree when entering a new tab and the tree was previously open
-vim.g.nvim_tree_auto_resize = 0 -- by default, will resize the tree to its saved width when opening a file
-vim.g.nvim_tree_disable_netrw = 1 -- by default, disables netrw
-vim.g.nvim_tree_hijack_netrw = 1 -- by default, prevents netrw from automatically opening when opening directories (but lets you keep its other utilities)
-vim.g.nvim_tree_add_trailing = 0 -- by default, append a trailing slash to folder names
-vim.g.nvim_tree_group_empty = 0 -- by default, compact folders that only contain a single folder into one node in the file tree
-vim.g.nvim_tree_lsp_diagnostics = 0 -- by default, will show lsp diagnostics in the signcolumn. See :help nvim_tree_lsp_diagnostics
-vim.g.nvim_tree_disable_window_picker = 1 -- by default, will disable the window picker.
-vim.g.nvim_tree_hijack_cursor = 0 -- by default, when moving cursor in the tree, will position the cursor at the start of the file on the current line
-vim.g.nvim_tree_icon_padding = ' ' -- space by default, used for rendering the space between the icon and the filename. Use with caution, it could break rendering if you set an empty string depending on your font.
-vim.g.nvim_tree_update_cwd = 0 -- by default, will update the tree cwd when changing nvim's directory (DirChanged event). Behaves strangely with autochdir set.
-vim.g.nvim_tree_window_picker_exclude = {
-    filetype = {
-        'packer',
-        'qf'
-    },
-    buftype = {
-        'terminal'
-    }
-}
--- Dictionary of buffer option names mapped to a list of option values that
--- indicates to the window picker that the buffer's window should not be
--- selectable.
-vim.g.nvim_tree_special_files = {} -- List of filenames that gets highlighted with NvimTreeSpecialFile
-vim.g.nvim_tree_show_icons = {
-    git = 1,
-    folders = 1,
-    files = 1,
-    folder_arrows = 1,
-}
--- If 0, do not show the icons for one of 'git' 'folder' and 'files'
--- 1 by default, notice that if 'files' is 1, it will only display
--- if nvim-web-devicons is installed and on your runtimepath.
--- if folder is 1, you can also tell folder_arrows 1 to show small arrows next to the folder icons.
--- but this will not work when you set indent_markers (because of UI conflict)
-
---  default will show icon by default if no icon is provided
---  default shows no icon by default
-vim.g.nvim_tree_icons = {
-    default = '',
-    symlink = '',
-    git = {
-        unstaged = "✗",
-        staged = "✓",
-        unmerged = "",
-        renamed = "➜",
-        untracked = "★",
-        deleted = "",
-        ignored = "◌"
-    },
-    folder = {
-        arrow_open = "",
-        arrow_closed = "",
-        default = "",
-        open = "",
-        empty = "",
-        empty_open = "",
-        symlink = "",
-        symlink_open = "",
-    },
-    lsp = {
-        hint = "",
-        info = "",
-        warning = "",
-        error = "",
-       }
-}
-
-vim.api.nvim_set_keymap('n', '<leader>e', ':NvimTreeToggle<CR>', { silent = true })
-vim.api.nvim_set_keymap('n', '<leader>r', ':NvimTreeRefresh<CR>', { silent = true })
-vim.api.nvim_set_keymap('n', '<leader>n', ':NvimTreeFindFile<CR>', { silent = true })
--- NvimTreeOpen and NvimTreeClose are also available if you need them
-
--- a list of groups can be found at `:help nvim_tree_highlight`
-vim.cmd('highlight NvimTreeFolderIcon guibg=blue')
-
--- Symbols.outline.nvim config
-vim.g.symbols_outline = {
-    highlight_hovered_item = true,
-    show_guides = true,
-    auto_preview = true,
-    position = 'right',
-    width = 25,
-    show_numbers = false,
-    show_relative_numbers = false,
-    show_symbol_details = true,
-    keymaps = { -- These keymaps can be a string or a table for multiple keys
-        close = {"<Esc>", "q"},
-        goto_location = "<Cr>",
-        focus_location = "o",
-        hover_symbol = "<C-space>",
-        rename_symbol = "r",
-        code_actions = "a",
-    },
-    lsp_blacklist = {},
-    symbol_blacklist = {},
-    symbols = {
-        File = {icon = "", hl = "TSURI"},
-        Module = {icon = "", hl = "TSNamespace"},
-        Namespace = {icon = "", hl = "TSNamespace"},
-        Package = {icon = "", hl = "TSNamespace"},
-        Class = {icon = "𝓒", hl = "TSType"},
-        Method = {icon = "ƒ", hl = "TSMethod"},
-        Property = {icon = "", hl = "TSMethod"},
-        Field = {icon = "", hl = "TSField"},
-        Constructor = {icon = "", hl = "TSConstructor"},
-        Enum = {icon = "ℰ", hl = "TSType"},
-        Interface = {icon = "ﰮ", hl = "TSType"},
-        Function = {icon = "", hl = "TSFunction"},
-        Variable = {icon = "", hl = "TSConstant"},
-        Constant = {icon = "", hl = "TSConstant"},
-        String = {icon = "𝓐", hl = "TSString"},
-        Number = {icon = "#", hl = "TSNumber"},
-        Boolean = {icon = "⊨", hl = "TSBoolean"},
-        Array = {icon = "", hl = "TSConstant"},
-        Object = {icon = "⦿", hl = "TSType"},
-        Key = {icon = "🔐", hl = "TSType"},
-        Null = {icon = "NULL", hl = "TSType"},
-        EnumMember = {icon = "", hl = "TSField"},
-        Struct = {icon = "𝓢", hl = "TSType"},
-        Event = {icon = "🗲", hl = "TSType"},
-        Operator = {icon = "+", hl = "TSOperator"},
-        TypeParameter = {icon = "𝙏", hl = "TSParameter"}
-    }
-}
-vim.api.nvim_set_keymap('n', '<leader>o', ':SymbolsOutline<CR>', { noremap = true, silent = true })
-
--- "Spaceline.nvim config
-vim.g.spaceline_seperate_style = 'none'
-
--- Nvim-Bufferline config
-require('bufferline').setup {
-  options = {
-    numbers = "none", -- "none" | "ordinal" | "buffer_id" | "both",
-    close_command = "bdelete! %d",       -- can be a string | function, see "Mouse actions"
-    right_mouse_command = "bdelete! %d", -- can be a string | function, see "Mouse actions"
-    left_mouse_command = "buffer %d",    -- can be a string | function, see "Mouse actions"
-    middle_mouse_command = nil,          -- can be a string | function, see "Mouse actions"
-    -- NOTE: this plugin is designed with this icon in mind,
-    -- and so changing this is NOT recommended, this is intended
-    -- as an escape hatch for people who cannot bear it for whatever reason
-    indicator_icon = '▎',
-    buffer_close_icon = '',
-    modified_icon = '●',
-    close_icon = '',
-    left_trunc_marker = '',
-    right_trunc_marker = '',
-    --- name_formatter can be used to change the buffer's label in the bufferline.
-    --- Please note some names can/will break the
-    --- bufferline so use this at your discretion knowing that it has
-    --- some limitations that will *NOT* be fixed.
-    name_formatter = function(buf)  -- buf contains a "name", "path" and "bufnr"
-      -- remove extension from markdown files for example
-      if buf.name:match('%.md') then
-        return vim.fn.fnamemodify(buf.name, ':t:r')
-      end
-    end,
-    max_name_length = 18,
-    max_prefix_length = 15, -- prefix used when a buffer is de-duplicated
-    tab_size = 18,
-    diagnostics = "nvim_lsp", -- false | "nvim_lsp",
-    diagnostics_indicator = function(count, level, diagnostics_dict, context)
-      return "("..count..")"
-    end,
-    -- NOTE: this will be called a lot so don't do any heavy processing here
-    custom_filter = function(buf_number)
-      -- filter out filetypes you don't want to see
-      if vim.bo[buf_number].filetype ~= "<i-dont-want-to-see-this>" then
-        return true
-      end
-      -- filter out by buffer name
-      if vim.fn.bufname(buf_number) ~= "<buffer-name-I-dont-want>" then
-        return true
-      end
-      -- filter out based on arbitrary rules
-      -- e.g. filter out vim wiki buffer from tabline in your work repo
-      if vim.fn.getcwd() == "<work-repo>" and vim.bo[buf_number].filetype ~= "wiki" then
-        return true
-      end
-    end,
-    offsets = {{filetype = "NvimTree", text = "File Explorer", text_align = "center" }, {filetype = "Outline", text = "Outline", text_align = "center" }},
-    show_buffer_icons = true, -- true | false, -- disable filetype icons for buffers
-    show_buffer_close_icons = true, -- true | false,
-    show_close_icon = true, -- true | false,
-    show_tab_indicators = true, -- true | false,
-    persist_buffer_sort = true, -- whether or not custom sorted buffers should persist
-    -- can also be a table containing 2 custom separators
-    -- [focused and unfocused]. eg: { '|', '|' }
-    separator_style = "thin", -- "slant" | "thick" | "thin" | { 'any', 'any' },
-    enforce_regular_tabs = false, -- false | true,
-    always_show_bufferline = true, -- true | false,
-    sort_by = 'id', -- 'id' | 'extension' | 'relative_directory' | 'directory' | 'tabs' | function(buffer_a, buffer_b)
-    -- add custom logic
-    --   return buffer_a.modified > buffer_b.modified
-    -- end
-  }
-}
-
--- Dashcoard-nvim config
-vim.g.dashboard_custom_header = {
-    '',
-    '',
-    '',
-    ' ███╗   ██╗ ███████╗ ██████╗  ██╗   ██╗ ██╗ ███╗   ███╗',
-    ' ████╗  ██║ ██╔════╝██╔═══██╗ ██║   ██║ ██║ ████╗ ████║',
-    ' ██╔██╗ ██║ █████╗  ██║   ██║ ██║   ██║ ██║ ██╔████╔██║',
-    ' ██║╚██╗██║ ██╔══╝  ██║   ██║ ╚██╗ ██╔╝ ██║ ██║╚██╔╝██║',
-    ' ██║ ╚████║ ███████╗╚██████╔╝  ╚████╔╝  ██║ ██║ ╚═╝ ██║',
-    ' ╚═╝  ╚═══╝ ╚══════╝ ╚═════╝    ╚═══╝   ╚═╝ ╚═╝     ╚═╝',
-    '',
-    '',
-}
-vim.g.dashboard_custom_shortcut= {
-    last_session = 'SPC s l',
-    find_history = 'SPC f h',
-    find_file = 'SPC f f',
-    new_file = 'SPC c n',
-    change_colorscheme =  'SPC t c',
-    find_word = 'SPC f a',
-    book_marks = 'SPC f b',
-}
-vim.g.dashboard_default_executive = 'telescope'
-vim.api.nvim_set_keymap('n', '<leader>ss', ':<C-u>SessionSave<CR>', { noremap = true, silent = true })
-vim.api.nvim_set_keymap('n', '<leader>sl', ':<C-u>SessionLoad<CR>', { noremap = true, silent = true })
-vim.api.nvim_set_keymap('n', '<leader>fh', ':<C-u>DashboardFindHistory<CR>', { noremap = true, silent = true })
-vim.api.nvim_set_keymap('n', '<leader>ff', ':<C-u>DashboardFindFile<CR>', { noremap = true, silent = true })
-vim.api.nvim_set_keymap('n', '<leader>tc', ':<C-u>DashboardChangeColorscheme<CR>', { noremap = true, silent = true })
-vim.api.nvim_set_keymap('n', '<leader>fa', ':<C-u>DashboardFindWord<CR>', { noremap = true, silent = true })
-vim.api.nvim_set_keymap('n', '<leader>fb', ':<C-u>DashboardJumpMark<CR>', { noremap = true, silent = true })
-vim.api.nvim_set_keymap('n', '<leader>cn', ':<C-u>DashboardNewFile<CR>', { noremap = true, silent = true })
-
--- Gitsings.nvim config
-require('gitsigns').setup {
-  signs = {
-    add          = {hl = 'GitSignsAdd'   , text = '│', numhl='GitSignsAddNr'   , linehl='GitSignsAddLn'},
-    change       = {hl = 'GitSignsChange', text = '│', numhl='GitSignsChangeNr', linehl='GitSignsChangeLn'},
-    delete       = {hl = 'GitSignsDelete', text = '_', numhl='GitSignsDeleteNr', linehl='GitSignsDeleteLn'},
-    topdelete    = {hl = 'GitSignsDelete', text = '‾', numhl='GitSignsDeleteNr', linehl='GitSignsDeleteLn'},
-    changedelete = {hl = 'GitSignsChange', text = '~', numhl='GitSignsChangeNr', linehl='GitSignsChangeLn'},
-  },
-  numhl = false,
-  linehl = false,
-  keymaps = {
-    -- Default keymap options
-    noremap = true,
-
-    ['n ]c'] = { expr = true, "&diff ? ']c' : '<cmd>lua require\"gitsigns.actions\".next_hunk()<CR>'"},
-    ['n [c'] = { expr = true, "&diff ? '[c' : '<cmd>lua require\"gitsigns.actions\".prev_hunk()<CR>'"},
-
-    ['n <leader>hs'] = '<cmd>lua require"gitsigns".stage_hunk()<CR>',
-    ['v <leader>hs'] = '<cmd>lua require"gitsigns".stage_hunk({vim.fn.line("."), vim.fn.line("v")})<CR>',
-    ['n <leader>hu'] = '<cmd>lua require"gitsigns".undo_stage_hunk()<CR>',
-    ['n <leader>hr'] = '<cmd>lua require"gitsigns".reset_hunk()<CR>',
-    ['v <leader>hr'] = '<cmd>lua require"gitsigns".reset_hunk({vim.fn.line("."), vim.fn.line("v")})<CR>',
-    ['n <leader>hR'] = '<cmd>lua require"gitsigns".reset_buffer()<CR>',
-    ['n <leader>hp'] = '<cmd>lua require"gitsigns".preview_hunk()<CR>',
-    ['n <leader>hb'] = '<cmd>lua require"gitsigns".blame_line(true)<CR>',
-
-    -- Text objects
-    ['o ih'] = ':<C-U>lua require"gitsigns.actions".select_hunk()<CR>',
-    ['x ih'] = ':<C-U>lua require"gitsigns.actions".select_hunk()<CR>'
-  },
-  watch_index = {
-    interval = 1000,
-    follow_files = true
-  },
-  current_line_blame = false,
-  current_line_blame_opts = {
-    delay = 1000,
-    virt_text_pos = 'eol'
-  },
-  sign_priority = 6,
-  update_debounce = 100,
-  status_formatter = nil, -- Use default
-  word_diff = false,
-  use_internal_diff = true,  -- If luajit is present
-}
-
--- Nvim-cololizer.lua config
-require'colorizer'.setup()
-
--- Minimap.vim config
-vim.g.minimap_width = 10
-vim.g.minimap_auto_start = 0
-vim.g.minimap_auto_start_win_enter = 0
-vim.g.minimap_highlight_search = 0
-vim.g.minimap_highlight_range = 0
-vim.g.minimap_block_filetypes = { 'terminal', 'NvimTree', 'Outline' }
-vim.g.minimap_close_filetypes = { 'dashboard', 'packer', }
