+++ nvim/lua/config/interface/nvim-treesitter-context.lua
@@ -1,0 +1,4 @@
+require'treesitter-context'.setup {
+    enable = true, -- Enable this plugin (Can be enabled/disabled later via commands)
+    throttle = true, -- Throttles plugin updates (may improve performance)
+}
