--- nvim/lua/config/tools.lua
@@ -1,358 +1,0 @@
-local vim = vim
-
--- Telescope.nvim config
-require('telescope').setup{
-  defaults = {
-    vimgrep_arguments = {
-      'rg',
-      '--color=never',
-      '--no-heading',
-      '--with-filename',
-      '--line-number',
-      '--column',
-      '--smart-case'
-    },
-    prompt_prefix = "> ",
-    selection_caret = "> ",
-    entry_prefix = "  ",
-    initial_mode = "insert",
-    selection_strategy = "reset",
-    sorting_strategy = "descending",
-    layout_strategy = "horizontal",
-    layout_config = {
-      horizontal = {
-        mirror = false,
-      },
-      vertical = {
-        mirror = false,
-      },
-    },
-    file_sorter =  require'telescope.sorters'.get_fuzzy_file,
-    file_ignore_patterns = {},
-    generic_sorter =  require'telescope.sorters'.get_generic_fuzzy_sorter,
-    winblend = 0,
-    border = {},
-    borderchars = { '─', '│', '─', '│', '╭', '╮', '╯', '╰' },
-    color_devicons = true,
-    use_less = true,
-    path_display = {},
-    set_env = { ['COLORTERM'] = 'truecolor' }, -- default = nil,
-    file_previewer = require'telescope.previewers'.vim_buffer_cat.new,
-    grep_previewer = require'telescope.previewers'.vim_buffer_vimgrep.new,
-    qflist_previewer = require'telescope.previewers'.vim_buffer_qflist.new,
-
-    -- Developer configurations: Not meant for general override
-    buffer_previewer_maker = require'telescope.previewers'.buffer_previewer_maker
-  },
-  extensions = {
-    fzf = {
-      fuzzy = true,                    -- false will only do exact matching
-      override_generic_sorter = false, -- override the generic sorter
-      override_file_sorter = true,     -- override the file sorter
-      case_mode = "smart_case",        -- or "ignore_case" or "respect_case"
-                                       -- the default case_mode is "smart_case"
-    }
-  }
-}
--- To get fzf loaded and working with telescope, you need to call
--- load_extension, somewhere after setup function:
-require('telescope').load_extension('fzf')
-
--- Telescope-project.nvim
-require'telescope'.load_extension('project')
-
--- Nvim-autopairs config
-require('nvim-autopairs').setup()
-
--- Nvim-ts-autotag config
-require'nvim-treesitter.configs'.setup {
-  autotag = {
-    enable = true,
-  }
-}
-vim.lsp.handlers['textDocument/publishDiagnostics'] = vim.lsp.with(
-    vim.lsp.diagnostic.on_publish_diagnostics,
-    {
-        underline = true,
-        virtual_text = {
-            spacing = 5,
-            severity_limit = 'Warning',
-        },
-        update_in_insert = true,
-    }
-)
-
--- Lazygit.nvim config
--- setup mapping to call :LazyGit
-vim.api.nvim_set_keymap('n', '<leader>lg', ':LazyGit<CR>', { noremap = true, silent = true })
-
--- VimTeX config
-vim.g.tex_flavor = 'latex'
-vim.g.vimtex_quickfix_mode = 0
-if vim.fn.has('unix') == 1 then
-    vim.g.vimtex_view_general_viewer = 'zathura'
-    vim.g.vimtex_view_method = 'zathura'
-end
-vim.g.vimtex_compiler_latexmk_engines = { _ = '-xelatex'}
-vim.g.vimtex_compiler_latexrun_engines ={ _ = 'xelatex'}
-vim.g.vimtex_compiler_progname = 'nvr'
-vim.g.vimtex_toc_config = {
-  name = 'TOC',
-  layers = {'content', 'todo', 'include'},
-  split_width = 25,
-  todo_sorted = 0,
-  show_help = 1,
-  show_numbers = 1,
-}
-
--- MarkdownPreview.nvim
--- set to 1, nvim will open the preview window after entering the markdown buffer
--- default: 0
-vim.g.mkdp_auto_start = 0
-
--- set to 1, the nvim will auto close current preview window when change
--- from markdown buffer to another buffer
--- default: 1
-vim.g.mkdp_auto_close = 1
-
--- set to 1, the vim will refresh markdown when save the buffer or
--- leave from insert mode, default 0 is auto refresh markdown as you edit or
--- move the cursor
--- default: 0
-vim.g.mkdp_refresh_slow = 0
-
--- set to 1, the MarkdownPreview command can be use for all files,
--- by default it can be use in markdown file
--- default: 0
-vim.g.mkdp_command_for_global = 0
-
--- set to 1, preview server available to others in your network
--- by default, the server listens on localhost (127.0.0.1)
--- default: 0
-vim.g.mkdp_open_to_the_world = 0
-
--- use custom IP to open preview page
--- useful when you work in remote vim and preview on local browser
--- more detail see: https://github.com/iamcco/markdown-preview.nvim/pull/9
--- default empty
-vim.g.mkdp_open_ip = ''
-
--- specify browser to open preview page
--- default: ''
-vim.g.mkdp_browser = 'chromium'
-
--- set to 1, echo preview page url in command line when open preview page
--- default is 0
-vim.g.mkdp_echo_preview_url = 0
-
--- a custom vim function name to open preview page
--- this function will receive url as param
--- default is empty
-vim.g.mkdp_browserfunc = ''
-
--- options for markdown render
--- mkit: markdown-it options for render
--- katex: katex options for math
--- uml: markdown-it-plantuml options
--- maid: mermaid options
--- disable_sync_scroll: if disable sync scroll, default 0
--- sync_scroll_type: 'middle', 'top' or 'relative', default value is 'middle'
---   middle: mean the cursor position alway show at the middle of the preview page
---   top: mean the vim top viewport alway show at the top of the preview page
---   relative: mean the cursor position alway show at the relative positon of the preview page
--- hide_yaml_meta: if hide yaml metadata, default is 1
--- sequence_diagrams: js-sequence-diagrams options
--- content_editable: if enable content editable for preview page, default: v:false
--- disable_filename: if disable filename header for preview page, default: 0
-vim.g.mkdp_preview_options = {
-      mkit = {},
-      katex = {},
-      uml = {},
-      maid = {},
-      disable_sync_scroll = 0,
-      sync_scroll_type = 'middle',
-      hide_yaml_meta = 1,
-      sequence_diagrams = {},
-      flowchart_diagrams = {},
-      content_editable = 'v:false',
-      disable_filename = 0
-      }
-
--- use a custom markdown style must be absolute path
--- like '/Users/username/markdown.css' or expand('~/markdown.css')
-vim.g.mkdp_markdown_css = ''
-
--- use a custom highlight style must absolute path
--- like '/Users/username/highlight.css' or expand('~/highlight.css')
-vim.g.mkdp_highlight_css = ''
-
--- use a custom port to start server or random for empty
-vim.g.mkdp_port = ''
-
--- preview page title
--- ${name} will be replace with the file name
-vim.g.mkdp_page_title = '「${name}」'
-
--- recognized filetypes
--- these filetypes will have MarkdownPreview... commands
-vim.g.mkdp_filetypes = {'markdown'}
-
-vim.api.nvim_set_keymap('n', '<leader>md', ':MarkdownPreviewToggle<CR>', { noremap = true, silent = true })
-
--- Diffview.nvim config
-local cb = require'diffview.config'.diffview_callback
-
-require'diffview'.setup {
-  diff_binaries = false,    -- Show diffs for binaries
-  file_panel = {
-    width = 35,
-    use_icons = true        -- Requires nvim-web-devicons
-  },
-  key_bindings = {
-    disable_defaults = false,                   -- Disable the default key bindings The `view` bindings are active in the diff buffers, only when the current
-    -- tabpage is a Diffview.
-    view = {
-      ["<tab>"]     = cb("select_next_entry"),  -- Open the diff for the next file 
-      ["<s-tab>"]   = cb("select_prev_entry"),  -- Open the diff for the previous file
-      ["<leader>e"] = cb("focus_files"),        -- Bring focus to the files panel
-      ["<leader>b"] = cb("toggle_files"),       -- Toggle the files panel.
-    },
-    file_panel = {
-      ["j"]             = cb("next_entry"),         -- Bring the cursor to the next file entry
-      ["<down>"]        = cb("next_entry"),
-      ["k"]             = cb("prev_entry"),         -- Bring the cursor to the previous file entry.
-      ["<up>"]          = cb("prev_entry"),
-      ["<cr>"]          = cb("select_entry"),       -- Open the diff for the selected entry.
-      ["o"]             = cb("select_entry"),
-      ["<2-LeftMouse>"] = cb("select_entry"),
-      ["-"]             = cb("toggle_stage_entry"), -- Stage / unstage the selected entry.
-      ["S"]             = cb("stage_all"),          -- Stage all entries.
-      ["U"]             = cb("unstage_all"),        -- Unstage all entries.
-      ["X"]             = cb("restore_entry"),      -- Restore entry to the state on the left side.
-      ["R"]             = cb("refresh_files"),      -- Update stats and entries in the file list.
-      ["<tab>"]         = cb("select_next_entry"),
-      ["<s-tab>"]       = cb("select_prev_entry"),
-      ["<leader>e"]     = cb("focus_files"),
-      ["<leader>b"]     = cb("toggle_files"),
-    }
-  }
-}
-
--- Nvim-toggleterm.lua
-require("toggleterm").setup{
-  -- size can be a number or function which is passed the current terminal
-  size = function(term)
-    if term.direction == "horizontal" then
-      return 15
-    elseif term.direction == "vertical" then
-      return vim.o.columns * 0.4
-    elseif term.direction == "float" then
-      return
-    end
-  end,
-  open_mapping = [[<c-\>]],
-  hide_numbers = true, -- hide the number column in toggleterm buffers
-  shade_filetypes = {},
-  shade_terminals = true,
-  shading_factor = '1', -- the degree by which to darken to terminal colour, default: 1 for dark backgrounds, 3 for light
-  start_in_insert = true,
-  insert_mappings = true, -- whether or not the open mapping applies in insert mode
-  persist_size = true,
-  direction = 'float', -- 'vertical' | 'horizontal' | 'window' | 'float',
-  close_on_exit = true, -- close the terminal window when the process exits
-  shell = vim.o.shell, -- change the default shell
-  -- This field is only relevant if direction is set to 'float'
-  float_opts = {
-    -- The border key is *almost* the same as 'nvim_win_open'
-    -- see :h nvim_win_open for details on borders however
-    -- the 'curved' border is a custom border type
-    -- not natively supported but implemented in this plugin.
-    border = 'curved', -- 'single' | 'double' | 'shadow' | 'curved' | ... other options supported by win open
-    width = vim.o.columns * 0.6,
-    height = 30,
-    winblend = 3,
-    highlights = {
-      border = "Normal",
-      background = "Normal",
-    }
-  }
-}
-
--- Nvim-treesitte-context config
-require'treesitter-context'.setup {
-    enable = true, -- Enable this plugin (Can be enabled/disabled later via commands)
-    throttle = true, -- Throttles plugin updates (may improve performance)
-}
-
--- Zen-mode.nvim config
-require'zen-mode'.setup {
-  window = {
-    backdrop = 0.95, -- shade the backdrop of the Zen window. Set to 1 to keep the same as Normal
-    -- height and width can be:
-    -- * an absolute number of cells when > 1
-    -- * a percentage of the width / height of the editor when <= 1
-    -- * a function that returns the width or the height
-    width = 120, -- width of the Zen window
-    height = 1, -- height of the Zen window
-    -- by default, no options are changed for the Zen window
-    -- uncomment any of the options below, or add other vim.wo options you want to apply
-    options = {
-      -- signcolumn = "no", -- disable signcolumn
-      -- number = false, -- disable number column
-      -- relativenumber = false, -- disable relative numbers
-      -- cursorline = false, -- disable cursorline
-      -- cursorcolumn = false, -- disable cursor column
-      -- foldcolumn = "0", -- disable fold column
-      -- list = false, -- disable whitespace characters
-    },
-  },
-  plugins = {
-    -- disable some global vim options (vim.o...)
-    -- comment the lines to not apply the options
-    options = {
-      enabled = true,
-      ruler = false, -- disables the ruler text in the cmd line area
-      showcmd = false, -- disables the command in the last line of the screen
-    },
-    twilight = { enabled = true }, -- enable to start Twilight when zen mode opens
-    gitsigns = { enabled = false }, -- disables git signs
-    tmux = { enabled = false }, -- disables the tmux statusline
-    -- this will change the font size on kitty when in zen mode
-    -- to make this work, you need to set the following kitty options:
-    -- - allow_remote_control socket-only
-    -- - listen_on unix:/tmp/kitty
-    kitty = {
-      enabled = false,
-      font = "+4", -- font size increment
-    },
-  },
-  -- callback where you can add custom code when the Zen window opens
-  on_open = function(win)
-  end,
-  -- callback where you can add custom code when the Zen window closes
-  on_close = function()
-  end,
-}
-
--- Nvim-comment config
-require('nvim_comment').setup {
-  -- Linters prefer comment and line to have a space in between markers
-  marker_padding = true,
-  -- should comment out empty or whitespace only lines
-  comment_empty = true,
-  -- Should key mappings be created
-  create_mappings = true,
-  -- Normal mode mapping left hand side
-  line_mapping = "gcc",
-  -- Visual/Operator mapping left hand side
-  operator_mapping = "gc",
-  -- Hook function to call before commenting takes place
-  hook = nil
-}
-
--- Nvim-lastplace config
-require'nvim-lastplace'.setup {
-    lastplace_ignore_buftype = {"quickfix", "nofile", "help", "dashboard", "NvimTree"},
-    lastplace_ignore_filetype = {"gitcommit", "gitrebase", "svn", "hgcommit"},
-    lastplace_open_folds = true
-}
