+++ nvim/lua/config/tools/lazygit.lua
@@ -1,0 +1,4 @@
+local vim = vim
+
+-- setup mapping to call :LazyGit
+vim.api.nvim_set_keymap('n', '<leader>lg', ':LazyGit<CR>', { noremap = true, silent = true })
